requirements.txt
boto3==1.36.7
Flask==3.1.0
python-dotenv==0.21.1
pydantic~=2.10.6
prometheus-flask-exporter~=0.22.0
gunicorn==23.0.0